my next rides service optimization

+++ flexigo/src/main/java/com/vektortelekom/android/vservice/ui/flexiride/FlexirideViewModel.kt
@@ -190,7 +190,7 @@ constructor(
             val additionalRiders = mutableListOf<Int>()
             for(person in personList) {
                 person.accountId?.let {
-                    additionalRiders.add(it)
+                    additionalRiders.add(it.toInt())
                 }
 
             }
