+++ flexigo/src/main/java/com/vektortelekom/android/vservice/ui/poolcar/reservation/PoolCarReservationViewModel.kt
@@ -263,7 +263,7 @@ constructor(private val poolCarRepository: PoolCarRepository,
                 val additionalRiders = mutableListOf<Int>()
                 for(person in personList) {
                     person.accountId?.let {
-                        additionalRiders.add(it)
+                        additionalRiders.add(it.toInt())
                     }
 
                 }
